requirements.txt
flask
pillow 